+++ lib/src/app.dart
@@ -11,6 +11,7 @@ import 'auth.dart';
 import 'bindings.dart' as js;
 import 'database.dart';
 import 'firestore.dart';
+import 'messaging.dart';
 
 /// Represents initialized Firebase application and provides access to the
 /// app's services.
@@ -40,6 +41,11 @@ class App {
   Firestore firestore() =>
       _firestore ??= new Firestore(nativeInstance.firestore());
   Firestore _firestore;
+  
+  /// Gets [Messaging] client for this application.
+  Messaging messaging() =>
+      _messaging ??= new Messaging(nativeInstance.messaging());
+  Messaging _messaging;
 
   /// Renders this app unusable and frees the resources of all associated
   /// services.